Added to HTML and CSS - arranged upper section and started second. Edited hero banner flexbox to better representt the prompt. Began adding elements of the second section.
NOTE: The prompt site seems to have the header and hero banner in-line.
Need to research how to keep those four flexboxes aligned. Additionally,
right margin of the sample image is off from left of screen. Need to research
why the .signup container is not equally spaced on all sides.